Renamed DataString to DataTagString. Added the DataTagCompound  to have like a map thing for the data tag system.

+++ src/main/java/at/korti/katlonengine/io/datatag/DataTagBase.java
@@ -7,7 +7,7 @@ import java.io.*;
  */
 public abstract class DataTagBase {
 
-    public static DataTagBase getType(byte id) {
+    public static DataTagBase getType(int id) {
         switch (id) {
             default:
                 return null;
